requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
python-dotenv==1.0.1

serpapi # New library for Google search
google-generativeai
openai>=1.52.0